requirements.txt
flask
opencensus-ext-azure